debug

+++ shared/src/main/scala/com/neo/sk/medusa/snake/package.scala
@@ -113,7 +113,7 @@ package object snake {
     header: Point = Point(20, 20),
     lastHeader: Point = Point(20, 20),
     direction: Point = Point(1, 0),
-    speed: Int = 10,
+    speed: Double = 10,
     speedUp : Double = 0.0,
     freeFrame : Int = 0,
     length: Int = 50,
